Fix #6, thanks @skironDotNet for the 1st issue
Using "Bad" is quite offensive. Coding styling is not for Bad or Good,
it s for a convenient purpose. As mentioned in #variable-names, I have
to deal with some not-very-powerful editor (geany, nano,...), that is
because I do not know how to use vi, vim, emnacs. Having a clear
indicator of object names is better Imho.

Maybe I extend the style to support multiple naming conventions.
At the end, it is all abount the consistency in the code...